requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
idna==3.4
pycparser==2.21
PyJWT==2.6.0
requests==2.28.1
urllib3==1.26.13
